Added true control over channel assignments
Removed a large chunk of code that created a function called idR, which was a very clunky way to assign unique channel IDs, and wasn't even truly user assignable. New version is, and has a large reduction in overall macro length as a result. Also fixed an issue where opening the first image in a given channel was tied to that channel's location in the input directory's file list (i.e. sorted by file name, which is likely to vary between users).