Welcome to v0.2!
Major changes are afoot.
 - Added color blind accommodations (on by default)
 - Simplified/reorganized UI
 - Reduced parameters required from user; many previously required inputs can now be left blank if unused or inferred from other provided information. -> Leaving a channel ID blank indicates that channel is not used, leaving a 0 in start or end Z clip location will default to first frame if start and last if end, and a 0 in both turns clipping off, otherwise provided defaults will be used
 - Repaired single channel input, which had broken while innovating multichannel handling
 - Added despeckling
 - Began adding user input min and max values for channels of interest, but this is on hold until I can make the UI better
 - Reduced redundancy - removed reused lines from within conditionals and placed them outside, etc.

I'm sure there're other things I'm forgetting -- This was a whirlwind overhaul that I got sort of lost in and didn't do the best job keeping track of the tweaks being made. Fundamentally, it still does everything it used to (and more), just more efficiently (programmatically and UI/UX wise). Clearly there's still more to do (see To Do list at the very top of the macro), but this is a big enough step forward that the version number will be incremented