Added option to specify save location if desired
Made this change after going to use this macro on non-local data for the first time and realizing that piping data from our NAS to FIJI to be processed and then back to the NAS to be saved makes much less sense than saving locally. When data, processing, and storage are all local (i.e. not via NAS), this option is not as important. Either option creates a new folder in the specified location, so the user does not need to create an output folder